pyproject.toml
[tool.poetry]
name = "tickercollector"
version = "0.1.0"
description = ""
authors = ["studio3104 <studio3104.com@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
bitflyer = {git = "https://github.com/RainMakerProject/bitFlyer.git", rev = "main"}
fire = "^0.4.0"
python-dotenv = "^0.19.2"
chart-handler = {git = "https://github.com/RainMakerProject/chartHandler.git", rev = "main"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
